% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchcorrection.R
\name{batch_detection}
\alias{batch_detection}
\title{batch_detection}
\usage{
batch_detection(mergeset, batch)
}
\arguments{
\item{mergeset}{A matrix of merged expression sets (rows = genes,
columns = samples).}

\item{batch}{Takes the results from \code{create_batch()} as input.}
}
\description{
Helper function to detect batches
}
