% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchcorrection.R
\name{batchCorrection_}
\alias{batchCorrection_}
\title{batchCorrection_}
\usage{
batchCorrection_(mergeset, batch)
}
\arguments{
\item{mergeset}{An ExpressionSet resulting from the \code{mergeGEO::mergeGEO()} function.}

\item{batch}{Takes the results from \code{createBatch_()} as input.}
}
\description{
Helper function to detect batches
}
