% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_knn_classifier.R
\name{build_predictive_knn}
\alias{build_predictive_knn}
\title{build_predictive_knn}
\usage{
build_predictive_knn(train_x, train_y, fit_cv, neighbors)
}
\arguments{
\item{train_x}{The learning data values.}

\item{train_y}{The learning data classes.}

\item{fit_cv}{Options for the cross validation}

\item{neighbors}{Compared neighbors for knn}
}
\description{
Function builds a knn classifier model based on the given data.
}
