% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_svm_classifier.R
\name{build_predictive_svm}
\alias{build_predictive_svm}
\title{build_predictive_svm}
\usage{
build_predictive_svm(train_x, train_y, fit_cv)
}
\arguments{
\item{train_x}{The learning data values.}

\item{train_y}{The learning data classes.}

\item{fit_cv}{Options for the cross validation}
}
\description{
Function builds a svm classifier model based on the given data.
}
