% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_diagnostic_models.R
\name{compare_diagnostic_models}
\alias{compare_diagnostic_models}
\title{compare_diagnostic_models}
\usage{
compare_diagnostic_models(modellist)
}
\arguments{
\item{modellist}{A list object containing a list of diagnostic models.
The output of the function `sigidentDiagnostic()`.}
}
\description{
Print an overview of all diagnostic models included
  in `modellist`.
}
