% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchcorrection.R
\name{create_batch}
\alias{create_batch}
\title{create_batch}
\usage{
create_batch(sample_metadata, study_metadata)
}
\arguments{
\item{sample_metadata}{A data frame. The data frame holding the
sample metadata.}

\item{study_metadata}{A data frame. The data frame holding the study
metadata.}
}
\description{
Helper function to create batch
}
