% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchcorrection.R
\name{create_diagnosisdesignbatch}
\alias{create_diagnosisdesignbatch}
\title{create_diagnosisdesignbatch}
\usage{
create_diagnosisdesignbatch(sample_metadata, study_metadata, controlname,
  targetname, targetcol)
}
\arguments{
\item{sample_metadata}{A data frame. The data frame holding the
sample metadata.}

\item{study_metadata}{A data frame. The data frame holding the study
metadata.}

\item{controlname}{A character string. Name of the the controls,
specified in the 'target' column of `sample_metadata`.}

\item{targetname}{A character string. Name of the the targets, specified
in the 'target' column of `sample_metadata`.}

\item{targetcol}{A character string. Columname of `sample_metadata`
holding the targets. Default: "target". Caution: this should not
be changed.}
}
\description{
Helper function to create diagnosis, design and batch
}
