% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchcorrection.R
\name{createBatch_}
\alias{createBatch_}
\title{createBatch_}
\usage{
createBatch_(sampleMetadata, studyMetadata)
}
\arguments{
\item{sampleMetadata}{A data frame. The data frame holding the sample metadata (output of the function \code{mergeGEO::readSampleMetadata_()}).}

\item{studyMetadata}{A data frame. The data frame holding the study metadata (output of the function \code{mergeGEO::readStudyMetadata_()}).}
}
\description{
Helper function to create batch
}
