% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{createBatchPlot_}
\alias{createBatchPlot_}
\title{createBatchPlot_}
\usage{
createBatchPlot_(correction_obj, filename = NULL, time)
}
\arguments{
\item{correction_obj}{An object. The output of the function `batchCorrection_()`.}

\item{filename}{A character string indicating the filename. If default (\code{NULL}) a plot named `PCplot{time}.png` will be created
inside the directory "./plots".}

\item{time}{A character string indicating if the plot is "before" or "after" batch correction. This information is integrated
into the filename.}
}
\description{
Helper function to create batchplot
}
