% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherfunctions.R
\name{createCombat_}
\alias{createCombat_}
\title{createCombat_}
\usage{
createCombat_(mergedset, batch, design)
}
\arguments{
\item{mergedset}{A large expression set. The output of the function `merge_()`.}

\item{batch}{Takes the results from \code{createBatch_()} as input.}

\item{design}{A object. The output of the function `createDiagnosisDesign_()`.}
}
\description{
Helper function to create a matrix with genes as rows and samples as columns
}
