% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{createDEGheatmap_}
\alias{createDEGheatmap_}
\title{createDEGheatmap_}
\usage{
createDEGheatmap_(mergeset, genes, patientcolors, filename = NULL)
}
\arguments{
\item{mergeset}{An ExpressionSet resulting from the \code{mergeGEO::mergeGEO()} function.}

\item{genes}{A object. The output of the function `identify.DEGs_()`.}

\item{filename}{A character string indicating the filename. If default (\code{NULL}) a plot named `DEG_heatmap.png` will be created
inside the directory "./plots".}
}
\description{
Helper function to create DEG heatmap
}
