% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchcorrection.R
\name{createDiagnosisDesign_}
\alias{createDiagnosisDesign_}
\title{createDiagnosisDesign_}
\usage{
createDiagnosisDesign_(sampleMetadata, studyMetadata, controlname,
  targetname, targetcol)
}
\arguments{
\item{sampleMetadata}{A data frame. The data frame holding the sample metadata (output of the function \code{mergeGEO::readSampleMetadata_()}).}

\item{studyMetadata}{A data frame. The data frame holding the study metadata (output of the function \code{mergeGEO::readStudyMetadata_()}).}

\item{controlname}{A character string. Name of the the controls, specified in the 'target' column of `sampleMetadata`.}

\item{targetname}{A character string. Name of the the targets, specified in the 'target' column of `sampleMetadata`.}

\item{targetcol}{A character string. Columname of `sampleMetadata` holding the targets. Default: "target". Caution: this should not be changed.}
}
\description{
Helper function to create diagnosis and design
}
