% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{createEnrichtedBarplot_}
\alias{createEnrichtedBarplot_}
\title{createEnrichtedBarplot_}
\usage{
createEnrichtedBarplot_(enrichmentobj, type, filename = NULL,
  showCategory = 20)
}
\arguments{
\item{enrichmentobj}{An object resulting from the function `goEnrichmentAnalysis_()`.}

\item{type}{A character string. One of eiter "GO" or "KEGG".}

\item{filename}{A character string indicating the filename. If default (\code{NULL}) a plot named `Enriched_{type}.png` will be created
inside the directory "./plots".}

\item{showCategory}{An integer. Indicating the number of maximum categories to show in barplot.}
}
\description{
Helper function to create enrichted barplots
}
