% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{createGridModelPlot_}
\alias{createGridModelPlot_}
\title{createGridModelPlot_}
\usage{
createGridModelPlot_(model, filename)
}
\arguments{
\item{model}{An object containing the model information. The ouput of the function `glmnetSignature_()` with \code{type = "grid"}.}

\item{filename}{A character string. The filename.}
}
\description{
Helper function to create the grid-model plot
}
