% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{createGridVarImpPlot_}
\alias{createGridVarImpPlot_}
\title{createGridVarImpPlot_}
\usage{
createGridVarImpPlot_(model, filename)
}
\arguments{
\item{model}{An object containing the model information. The ouput of the function `glmnetSignature_()` with \code{type = "grid"}.}

\item{filename}{A character string. The filename.}
}
\description{
Helper function to create the variable importance plot
}
