% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{createImportBoxplot_}
\alias{createImportBoxplot_}
\title{createImportBoxplot_}
\usage{
createImportBoxplot_(mergeset, filename = NULL)
}
\arguments{
\item{mergeset}{A merged ExpressionSet.}

\item{filename}{A character string indicating the filename. If default (\code{NULL}) a plot named `import_boxplot.png` will be created
inside the directory "./plots".}
}
\description{
Helper function to create boxplot of mergeset
}
