% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{createROCplot_}
\alias{createROCplot_}
\title{createROCplot_}
\usage{
createROCplot_(roc, filename)
}
\arguments{
\item{roc}{An object containing the roc information. The ouput of the function `glmnetSignature_()`
with \code{type = "elastic"} or \code{type = "lasso"}.}

\item{filename}{A character string. The filename.}
}
\description{
Helper function to create roc plots
}
