% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{createSurvPlot_}
\alias{createSurvPlot_}
\title{createSurvPlot_}
\usage{
createSurvPlot_(fit, RiskTable, filename)
}
\arguments{
\item{fit}{A cox proportional hazards model. The output of the function `fitKaplanEstimator_()` or `prognosticClassifier_()`.}

\item{RiskTable}{A data.frame. The output of the function `prognosticClassifier_()`.}

\item{filename}{A character string. The filename.}
}
\description{
Helper function to create survival plot
}
