% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherfunctions.R
\name{export_deg_annotations}
\alias{export_deg_annotations}
\title{export_deg_annotations}
\usage{
export_deg_annotations(mergedset, genes, idtype)
}
\arguments{
\item{genes}{A object. The output of the function `identify_degs()`.}
}
\description{
Helper function to export DEG annotations from mergedset.
}
