% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{extractGOterms_}
\alias{extractGOterms_}
\title{extractGOterms_}
\usage{
extractGOterms_(entrez, species, FDR = NULL)
}
\arguments{
\item{entrez}{A character vector containing the Entrez-IDs.}

\item{species}{A character string indicating the sample's species. Currently supported: "Hs".}

\item{FDR}{The false discovery rate passed to `limma::goana`.}
}
\description{
Helper function to extract GO terms
}
