% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_gbm.R
\name{gbm_classifier}
\alias{gbm_classifier}
\title{gbm_classifier}
\usage{
gbm_classifier(traininglist, seed, nfolds, repeats, tunelength)
}
\arguments{
\item{traininglist}{A list object containing the training data. The output
of the function `create_training_test_split()`.}

\item{seed}{Intilization state of random number generator}

\item{nfolds}{A integer. The number of folds used for cross validation
(default: 10).}

\item{repeats}{An integer. The number of repeated cross validations
(default = 5).}

\item{tunelength}{An integer. The caret tuning length (default = 10).}
}
\description{
Helper function to create a gbm classifier model
}
