% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_map_sig.R
\name{gene_map_sig}
\alias{gene_map_sig}
\title{gene_map_sig}
\usage{
gene_map_sig(mergeset, model)
}
\arguments{
\item{mergeset}{A matrix of merged expression sets (rows = genes,
columns = samples). The output of the funtion
`sigident.preproc::load_geo_data()`.}

\item{model}{An object containing the model information. The
ouput of the function `glmnetSignature_()` with \code{type = "grid"}.}
}
\description{
Helper function to map relevant input variables of a diagnostic
  model to corresponding IDs.
}
