% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_sig.R
\name{geneMapSig_}
\alias{geneMapSig_}
\title{geneMapSig_}
\usage{
geneMapSig_(mergeset, model)
}
\arguments{
\item{mergeset}{A matrix of merged expression sets (rows = genes, columns = samples).}

\item{model}{An object containing the model information. The ouput of the function `glmnetSignature_()` with \code{type = "grid"}.}
}
\description{
Helper function to map relevant input variables of a diagnostic model to corresponding IDs.
}
