% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prognostic_sig.R
\name{generate_expression_pattern}
\alias{generate_expression_pattern}
\title{generate_expression_pattern}
\usage{
generate_expression_pattern(
  classifier_studies,
  sig_cov,
  mergeset,
  sample_metadata
)
}
\arguments{
\item{classifier_studies}{A character vector. Names of the studies
used to train the classifier.}

\item{sig_cov}{A data.frame. Output of the function `univ_cox()`.}

\item{mergeset}{A matrix of merged expression sets (rows = genes,
columns = samples). The output of the funtion
`sigident.preproc::load_geo_data()`.}

\item{sample_metadata}{A data frame. The data frame holding the
sample metadata.}
}
\description{
Helper function to generate expression pattern
}
