% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prognostic_sig.R
\name{generateExpressionPattern_}
\alias{generateExpressionPattern_}
\title{generateExpressionPattern_}
\usage{
generateExpressionPattern_(classifier_studies, sigCov, mergeset,
  studyMetadata, sampleMetadata, controlname, targetname, targetcol)
}
\arguments{
\item{classifier_studies}{A character vector. Names of the studies used to train the classifier.}

\item{sigCov}{A data.frame. Output of the function `univCox_()`.}

\item{mergeset}{A matrix of merged expression sets (rows = genes, columns = samples).}
}
\description{
Helper function to generate expression pattern
}
