% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_diagnostic_models.R
\name{get_diagnostic_lambda_values}
\alias{get_diagnostic_lambda_values}
\title{get_diagnostic_lambda_values}
\usage{
get_diagnostic_lambda_values(modellist)
}
\arguments{
\item{modellist}{A list object containing a list of diagnostic models.
The output of the function `sigidentDiagnostic()`.}
}
\description{
Print an overview of all lambda values of the diagnostic models
  in `modellist`.
}
