% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prognostic_sig.R
\name{get_survival_time}
\alias{get_survival_time}
\title{get_survival_time}
\usage{
get_survival_time(
  sample_metadata,
  genes,
  idtype,
  discoverystudies_w_timedata,
  datadir
)
}
\arguments{
\item{sample_metadata}{A data frame. The data frame holding the
sample metadata.}

\item{genes}{A object. The output of the function
`sigident.func::identify_degs()`}

\item{idtype}{A character string. The type of ID used to name the
genes. One of 'entrez' or 'affy' intended to use either entrez IDs or
affy IDs. Caution: when using entrez IDs, missing and duplicated IDs
are being removed!}

\item{discoverystudies_w_timedata}{A list that contains specifications
on the study/studies that contain(s) survival time information.}

\item{datadir}{A character string. Path to the data-folder inside the
metadata folder.}
}
\description{
Helper function to get survival time.
}
