% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareDiagnosticModels.R
\name{getDiagnosticLambdaValues}
\alias{getDiagnosticLambdaValues}
\title{getDiagnosticLambdaValues}
\usage{
getDiagnosticLambdaValues(modellist)
}
\arguments{
\item{modellist}{A list object containing a list of diagnostic models. The output of the function `sigidentDiagnostic()`.}
}
\description{
Print an overview of all lambda values of the diagnostic models in `modellist`.
}
