% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prognostic_sig.R
\name{getSurvivalTime_}
\alias{getSurvivalTime_}
\title{getSurvivalTime_}
\usage{
getSurvivalTime_(studyMetadata, sampleMetadata, genes, idtype,
  discoverystudies.w.timedata, targetname, controlname, targetcol, datadir)
}
\arguments{
\item{genes}{A object. The output of the function `identifyDEGs_()`.}

\item{discoverystudies.w.timedata}{A list that contains specifications on the study/studies that contain(s) survival time information.}

\item{datadir}{A character string. Path to the data-folder inside the metadata folder.}
}
\description{
Helper function to get survival time.
}
