% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{goDiffReg_}
\alias{goDiffReg_}
\title{goDiffReg_}
\usage{
goDiffReg_(mergeset, design)
}
\arguments{
\item{mergeset}{An ExpressionSet resulting from the \code{mergeGEO::mergeGEO()} function.}

\item{design}{A object. The output of the function `createDiagnosisDesign_()`.}
}
\description{
Helper function to ... (TODO what do we do here?)
}
