% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{goEnrichmentAnalysis_}
\alias{goEnrichmentAnalysis_}
\title{goEnrichmentAnalysis_}
\usage{
goEnrichmentAnalysis_(entrez, OrgDB, organism, fitlm, pathwayid, species,
  plotdir = NULL)
}
\arguments{
\item{entrez}{A character vector containing the Entrez-IDs.}

\item{fitlm}{An object, returned by `goDiffReg_()`.}

\item{plotdir}{A character string. Path to the folder to store resulting plots. Default: "./plots/".}
}
\description{
Helper function to perform enrichment analysis
}
