% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG.R
\name{identify_degs}
\alias{identify_degs}
\title{identify_degs}
\usage{
identify_degs(mergeset, design, q_value = 0.01)
}
\arguments{
\item{mergeset}{A matrix of merged expression sets (rows = genes,
columns = samples).}

\item{design}{A object. The output of the function
`sigident.preproc::geo_create_diagnosisdesignbatch()`.}

\item{q_value}{A numeric value for the q-value (false discovery rate)
(default=0.01).}
}
\description{
Helper function to identify DEGs based on the limma package
}
