% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG.R
\name{limma_top_table}
\alias{limma_top_table}
\title{limma_top_table}
\usage{
limma_top_table(mergeset, design, q_value)
}
\arguments{
\item{mergeset}{A matrix of merged expression sets (rows = genes,
columns = samples).}

\item{design}{A object. The output of the function
`sigident.preproc::geo_create_diagnosisdesignbatch()`.}

\item{q_value}{A numeric value for the q-value (false discovery rate)
(default=0.01).}
}
\description{
Helper function to get DEG results
}
