% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG.R
\name{limmaTopTable_}
\alias{limmaTopTable_}
\title{limmaTopTable_}
\usage{
limmaTopTable_(mergeset, design, qValue)
}
\arguments{
\item{mergeset}{A matrix of merged expression sets (rows = genes, columns = samples).}

\item{design}{A object. The output of the function `createDiagnosisDesign_()`.}

\item{qValue}{A numeric value for the q-value (false discovery rate) (default=0.01).}
}
\description{
Helper function to get DEG results
}
