% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherfunctions.R
\name{merge_}
\alias{merge_}
\title{merge_}
\usage{
merge_(esetlist)
}
\arguments{
\item{esetlist}{A list of ExpressionSets with consistent phenoData to merge}
}
\description{
Helper function to merge ExpressionSets
}
