% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_cvplot}
\alias{plot_cvplot}
\title{plot_cvplot}
\usage{
plot_cvplot(cv_obj, filename)
}
\arguments{
\item{cv_obj}{A object. The result of the function `sigident_signature()`.}

\item{filename}{A character string. The filename.}
}
\description{
Helper function to create cross-validation plots
}
