% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_import_boxplot}
\alias{plot_import_boxplot}
\title{plot_import_boxplot}
\usage{
plot_import_boxplot(mergeset, filename = NULL)
}
\arguments{
\item{mergeset}{A matrix of merged expression sets (rows = genes, columns = samples).}

\item{filename}{A character string indicating the filename. If default
(\code{NULL}) a plot named `import_boxplot.png` will be created
inside the directory "./plots".}
}
\description{
Helper function to create boxplot of mergeset
}
