% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_rocplot}
\alias{plot_rocplot}
\title{plot_rocplot}
\usage{
plot_rocplot(roc, filename)
}
\arguments{
\item{roc}{An object containing the roc information.
The ouput of the function `glmnetSignature_()`
with \code{type = "elastic"} or \code{type = "lasso"}.}

\item{filename}{A character string. The filename.}
}
\description{
Helper function to create roc plots
}
