% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_rf.R
\name{predict_caret}
\alias{predict_caret}
\title{predict_caret}
\usage{
predict_caret(model, test_x)
}
\arguments{
\item{model}{The caret model.}

\item{test_x}{The data to be classified.}
}
\description{
Function to classify given data based on the created
  model.
}
