% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_knn_classifier.R
\name{predict_knn}
\alias{predict_knn}
\title{predict_knn}
\usage{
predict_knn(model, test_x)
}
\arguments{
\item{model}{A list object containing the training data. The output
of the function `build_predictive_knn()`.}

\item{test_x}{The data to be classified.}
}
\description{
Function to classify given data based on the created
  model.
}
