% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_rf.R
\name{predict_rf}
\alias{predict_rf}
\title{predict_rf}
\usage{
predict_rf(model, test_x)
}
\arguments{
\item{model}{A list object containing the training data. The output
of the function `build_predictive_rf()`.}

\item{test_x}{The data to be classified.}
}
\description{
Function to classify given data based on the created
  model.
}
