% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_svm.R
\name{predict_svm}
\alias{predict_svm}
\title{predict_svm}
\usage{
predict_svm(model, test_x)
}
\arguments{
\item{model}{A list object containing the training data. The output
of the function `build_predictive_svm()`.}

\item{test_x}{The data to be classified.}
}
\description{
Function to classify given data based on the created
  model.
}
