% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prognostic_sig.R
\name{prognosticClassifier_}
\alias{prognosticClassifier_}
\title{prognosticClassifier_}
\usage{
prognosticClassifier_(PatternCom, idtype, validationstudiesinfo, datadir,
  targetcol, targetname, controlname)
}
\arguments{
\item{PatternCom}{A data.frame. The output of the function `generateExpressionPattern_()`.}

\item{idtype}{A character string. The type of ID used to name the genes. One of 'entrez' or 'affy' intended to use either entrez IDs or
affy IDs. Caution: when using entrez IDs, missing and duplicated IDs are being removed!}

\item{validationstudiesinfo}{A list that contains specifications on the study that contains the validation information.}

\item{datadir}{A character string. Path to the data-folder inside the metadata folder.}

\item{targetcol}{A character string. Columname of `sampleMetadata` holding the targets. Default: "target". Caution: this should not be changed.}

\item{targetname}{A character string. Name of the the targets, specified in the 'target' column of `sampleMetadata`.}

\item{controlname}{A character string. Name of the the controls, specified in the 'target' column of `sampleMetadata`.}
}
\description{
Helper function to get survival time.
}
