% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG.R
\name{qSelection_}
\alias{qSelection_}
\title{qSelection_}
\usage{
qSelection_(sampleMetadata, deg.q.selection = NULL)
}
\arguments{
\item{sampleMetadata}{A data frame. The data frame holding the sample metadata (output of the function \code{mergeGEO::readSampleMetadata_()}).}

\item{deg.q.selection}{A numeric value between 0 and 1 indicating the desired q-Value during DEG analysis. Default: NULL, which means, that q is
calculated the following: \emph{1/length(mergeset@featureData@data$ID)}.}
}
\description{
Helper function to select qValues
}
