% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigident_signature.R
\name{sigident_signature}
\alias{sigident_signature}
\title{sigident_signature}
\usage{
sigident_signature(
  traininglist,
  type,
  a = NULL,
  nfolds = 10,
  repeats = 5,
  tunelength = 10,
  seed,
  ncores = 4
)
}
\arguments{
\item{traininglist}{A list object containing the training data. The output
of the function `create_training_test_split()`.}

\item{type}{A character string. The algorihm used to perform calculations.
Currently implemented are \emph{"glmnet", "lasso", "elastic",
"svm" (support vector machine), "rf" (random forest) and "knn" (k-nearest
neighbors)}.}

\item{a}{A numeric between 0 and 1. The elastic net mixing parameter 'alpha'
passed to `glmnet::glmnet()`.}

\item{nfolds}{A integer. The number of folds used for cross validation
(default: 10).}

\item{repeats}{An integer. The number of repeated cross validations
(default = 5).}

\item{tunelength}{An integer. The caret tuning length (default = 10).}

\item{seed}{A integer value. Seed to make machine learning algorithms
reproducible. Default: 111.}

\item{ncores}{An integers. The number of CPU cores used for running
the algorithms (default = 4).}
}
\description{
Helper function to identify a multi-gene classifier
}
