% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigidentDiagnostic.R
\name{sigidentDiagnostic}
\alias{sigidentDiagnostic}
\title{Perform Diagnostic Signature Analyses in Gene Expression Datasets
  Derived from MicroArrays}
\usage{
sigidentDiagnostic(
  mergeset,
  diagnosis,
  seed = 111,
  nfolds = 10,
  repeats = 5,
  tunelength = 10,
  split = 0.8,
  plotdir = paste0(tempdir(), "/plots/"),
  colindices = NULL
)
}
\arguments{
\item{mergeset}{A matrix of merged expression sets (rows = genes,
columns = samples). The output of the funtion
`sigident.preproc::load_geo_data()`.}

\item{diagnosis}{An object. The output of the function
`create_diagnosisdesignbatch()`.}

\item{seed}{A integer value. Seed to make machine learning algorithms
reproducible. Default: 111.}

\item{nfolds}{A integer. The number of folds used for cross validation
(default: 10).}

\item{repeats}{An integer. The number of repeated cross validations
(default = 5).}

\item{tunelength}{An integer. The caret tuning length (default = 10).}

\item{split}{A numeric value between 0 and 1. The proportion of the data
to be integrated into the training set for machine learning
(default: 0.8).}

\item{plotdir}{A character string. Path to the folder to store resulting
plots (default = paste0(tempdir(), "/plots/")).}

\item{colindices}{A vector of integers. Indices of columns to run the
validation on (default = NULL).}
}
\description{
One function to perform diagnostic signature analysis.
}
