% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigident.R
\name{sigidentDNA}
\alias{sigidentDNA}
\title{Perform Signature Analyses in Gene Expression Datasets}
\usage{
sigidentDNA(mergedset, plotdir, csvdir, targetcol, controlname, targetname,
  species, deg.q.selection = NULL, seed = 111, traintest.split = 0.8)
}
\description{

}
\examples{
\dontrun{
}

}
