% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigident.R
\name{sigidentMicroarray}
\alias{sigidentMicroarray}
\title{Perform Signature Analyses in Gene Expression Datasets Derived from MicroArrays}
\usage{
sigidentMicroarray(mergeset, controlname, targetname, studyMetadata,
  sampleMetadata, species, OrgDB, organism, pathwayid,
  deg.q.selection = NULL, seed = 111, nfolds = 10, split = 0.8,
  plotdir = "./plots/", csvdir = "./tables/", targetcol = "target")
}
\arguments{
\item{mergeset}{A matrix of merged expression sets (rows = genes, columns = samples).}

\item{controlname}{A character string. Name of the the controls, specified in the 'target' column of `sampleMetadata`.}

\item{targetname}{A character string. Name of the the targets, specified in the 'target' column of `sampleMetadata`.}

\item{studyMetadata}{A data frame. The data frame holding the study metadata.}

\item{sampleMetadata}{A data frame. The data frame holding the sample metadata.}

\item{species}{A character string indicating the sample's species. Currently supported: "Hs".}

\item{OrgDB}{A character string indicating the OrgDb. Currently supported: "org.Hs.eg.db".}

\item{organism}{A character string indicating the organism. Currently supported: "hsa".}

\item{pathwayid}{A character string indicating the pathway to show in the enrichment analysis. Currently supported: "hsa04110".}

\item{deg.q.selection}{A numeric value between 0 and 1 indicating the desired q-Value during DEG analysis. Default: NULL, which means, that q is
calculated the following: \emph{1/length(sample-IDs in discovery studies)}.}

\item{seed}{A integer value. Seed to make machine learning algorithms reproducible. Default: 111.}

\item{nfolds}{A integer. The number of folds used for cross validation. Default: 10.}

\item{split}{A numeric value between 0 and 1. The proportion of the data to be integrated into the training set for machine learning. Default: 0.8.}

\item{plotdir}{A character string. Path to the folder to store resulting plots. Default: "./plots/".}

\item{csvdir}{A character string. Path to the folder to store output tables. Default: "./tables/".}

\item{targetcol}{A character string. Columname of `sampleMetadata` holding the targets. Default: "target". Caution: this should not be changed.}
}
\description{
One function that wrappes all the functionality of the sigident package.
}
