% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigidentPrognostic.R
\name{sigidentPrognostic}
\alias{sigidentPrognostic}
\title{Perform Prognostic Signature Analyses in Gene Expression Datasets Derived from MicroArrays}
\usage{
sigidentPrognostic(mergeset, studyMetadata, sampleMetadata, idtype,
  discoverystudies.w.timedata, classifier_studies, validationstudiesinfo,
  targetname, controlname, datadir, plotdir = "./plots/",
  csvdir = "./tables/", targetcol = "target")
}
\arguments{
\item{mergeset}{A matrix of merged expression sets (rows = genes, columns = samples).}

\item{discoverystudies.w.timedata}{A list that contains specifications on the study/studies that contain(s) survival time information.}

\item{classifier_studies}{A character vector. Names of the studies used to train the classifier.}

\item{validationstudiesinfo}{A list that contains specifications on the study that contains the validation information.}

\item{datadir}{A character string. Path to the data-folder inside the metadata folder.}

\item{plotdir}{A character string. Path to the folder to store resulting plots. Default: "./plots/".}
}
\description{
One function to perform prognostic signature analysis.
}
