% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_diagnostic.R
\name{signature_}
\alias{signature_}
\title{signature_}
\usage{
signature_(traininglist, type, alpha = NULL, nfolds = 10, seed)
}
\arguments{
\item{traininglist}{A list object containint the training data. The output of the function `createTrainingTest_()`.}

\item{type}{A character string. The algorihm used to perform calculations. Currently implemented are \emph{"grid", "lasso", "elastic"} and
\emph{"ridge"}.}

\item{alpha}{A numeric between 0 and 1. The elasticnet mixing parameter passed to `glmnet::glmnet()`.}

\item{nfolds}{A integer. The number of folds used for cross validation. Default: 10.}

\item{seed}{A integer value. Seed to make machine learning algorithms reproducible. Default: 111.}
}
\description{
Helper function to conduct lasso or elastic net regualization for fitting a GLM in order to identify a multi-gene classifier
}
