% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_svm.R
\name{svm_classifier}
\alias{svm_classifier}
\title{svm_classifier}
\usage{
svm_classifier(traininglist, seed, nfolds, repeats)
}
\arguments{
\item{traininglist}{A list object containing the training data. The output
of the function `create_training_test_split()`.}

\item{seed}{Intilization state of random number generator}

\item{nfolds}{A integer. The number of folds used for cross validation
(default: 10).}

\item{repeats}{An integer. The number of repeated cross validations
(default = 3).}
}
\description{
Helper function to create a svm classifier model
}
