% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prognostic_sig.R
\name{univ_cox}
\alias{univ_cox}
\title{univ_cox}
\usage{
univ_cox(survtable, genes)
}
\arguments{
\item{survtable}{A data.frame. Output of the function `get_survival_time()`.}

\item{genes}{A object. The output of the function
`sigident.func::identify_degs()`}
}
\description{
Helper function to compute univariate cox regression and
  determine significance of each gene through separate univariate Cox
  regressions
}
