% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_sig.R
\name{validate_diagnostic_signature}
\alias{validate_diagnostic_signature}
\title{validate_diagnostic_signature}
\usage{
validate_diagnostic_signature(
  validationstudylist,
  models,
  genes,
  idtype,
  targetname,
  controlname,
  targetcol,
  datadir
)
}
\arguments{
\item{validationstudylist}{A list, containing metainformation on the
study used for validation of the diagnostic signature.}

\item{models}{A list of prediction models. Usually the output of the
function `sigidentDiagnostic`.}

\item{genes}{A object. The output of the function `identify_degs()`.}

\item{idtype}{A character string. The type of ID used to name the
genes. One of 'entrez' or 'affy' intended to use either entrez IDs or
affy IDs. Caution: when using entrez IDs, missing and duplicated IDs
are being removed!}

\item{targetname}{A character string. Name of the the targets, specified
in the 'target' column of `sample_metadata`.}

\item{controlname}{A character string. Name of the the controls,
specified in the 'target' column of `sample_metadata`.}

\item{targetcol}{A character string. Columname of `sample_metadata`
holding the targets. Default: "target". Caution: this should not
be changed.}

\item{datadir}{A character string. Path to the data-folder inside the
metadata folder.}
}
\description{
Helper function to validate diagnostic signatures
}
