% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_sig.R
\name{validate_diagnostic_signatures}
\alias{validate_diagnostic_signatures}
\title{validate_diagnostic_signatures}
\usage{
validate_diagnostic_signatures(
  validationstudylist,
  models,
  genes,
  idtype,
  datadir
)
}
\arguments{
\item{validationstudylist}{A list, containing metainformation on the
study used for validation of the diagnostic signature.}

\item{models}{A list of prediction models. Usually the output of the
function `sigidentDiagnostic`.}

\item{genes}{A object. The output of the function
`sigident.func::identify_degs()`}

\item{idtype}{A character string. The type of ID used to name the
genes. One of 'entrez' or 'affy' intended to use either entrez IDs or
affy IDs. Caution: when using entrez IDs, missing and duplicated IDs
are being removed!}

\item{datadir}{A character string. Path to the data-folder inside the
metadata folder.}
}
\description{
Helper function to validate diagnostic signatures
}
