% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_sig.R
\name{validateDiagnosticSignature_}
\alias{validateDiagnosticSignature_}
\title{validateDiagnosticSignature_}
\usage{
validateDiagnosticSignature_(validationstudylist, models, genes, idtype,
  targetname, controlname, targetcol, datadir)
}
\arguments{
\item{validationstudylist}{A list, containing metainformation on the study used for validation of the diagnostic signature.}

\item{models}{A prediction model to be used for validation.}

\item{genes}{A object. The output of the function `identifyDEGs_()`.}

\item{idtype}{A character string. The type of ID used to name the genes. One of 'entrez' or 'affy' intended to use either entrez IDs or
affy IDs. Caution: when using entrez IDs, missing and duplicated IDs are being removed!}

\item{targetname}{A character string. Name of the the targets, specified in the 'target' column of `sampleMetadata`.}

\item{controlname}{A character string. Name of the the controls, specified in the 'target' column of `sampleMetadata`.}

\item{targetcol}{A character string. Columname of `sampleMetadata` holding the targets. Default: "target". Caution: this should not be changed.}

\item{datadir}{A character string. Path to the data-folder inside the metadata folder.}
}
\description{
Helper function to validate diagnostic signatures
}
