% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchcorrection.R
\name{batchCorrection_}
\alias{batchCorrection_}
\title{batchCorrection_}
\usage{
batchCorrection_(mergedset, batch, design, idtype)
}
\arguments{
\item{mergedset}{An ExpressionSet. The output of the function `merge_()`.}

\item{batch}{Takes the results from \code{createBatch_()} as input.}

\item{design}{A object. The output of the function `createDiagnosisDesignBatch_()`.}
}
\description{
Helper function correcting for batch effects and mapping affy probes to Entrez IDs
}
\details{
This function takes a Bioconductor's ExpressionSet class (the output of the function `merge()`) and outputs a batch corrected
  matrix containing expression data. In order to correct for occurring batch effects and other unwanted variation in high-throughput
  experiments the `ComBat` function from the sva package is conducted.
  The affy probes are mapped to their Entrez IDs. Thereby, empty and replicated character strings are
  removed.
}
\references{
W.E. Johnson, C. Li, and A. Rabinovic. Adjusting batch effects in microarray data using empirical bayes methods. Biostatistics, 8(1):118–127, 2007.
  Jeffrey T. Leek, W. Evan Johnson, Hilary S. Parker, Elana J. Fertig, Andrew E. Jaffe, John D. Storey, Yuqing Zhang and Leonardo Collado Torres (2019). sva: Surrogate Variable Analysis. R package version 3.30.1.
}
