% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_knn.R
\name{build_predictive_knn}
\alias{build_predictive_knn}
\title{build_predictive_knn}
\usage{
build_predictive_knn(train_x, train_y, trn_ctrl)
}
\arguments{
\item{train_x}{The learning data values.}

\item{train_y}{The learning data classes.}

\item{trn_ctrl}{Options for the cross validation}
}
\description{
Function builds a knn classifier model based on the given data.
}
