% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_random_f_classifier.R
\name{build_predictive_rf}
\alias{build_predictive_rf}
\title{build_predictive_rf}
\usage{
build_predictive_rf(train_x, train_y, train_rf, metric)
}
\arguments{
\item{train_x}{The learning data values.}

\item{train_y}{The learning data classes.}

\item{fit_cv}{Options for the cross validation}
}
\description{
Function builds a random forest classifier model based on
the given data.
}
