% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_sig.R
\name{create_training_test_split}
\alias{create_training_test_split}
\title{create_training_test_split}
\usage{
create_training_test_split(diagnosis, mergeset, split = 0.8, seed = 111)
}
\arguments{
\item{diagnosis}{An object. The output of the function
`create_diagnosisdesignbatch()`.}

\item{mergeset}{A matrix of merged expression sets (rows = genes,
columns = samples). The output of the funtion
`sigident.preproc::load_geo_data()`.}

\item{split}{A numeric value between 0 and 1. The proportion of the data
to be integrated into the training set for machine learning
(default: 0.8).}

\item{seed}{A integer value. Seed to make machine learning algorithms
reproducible. Default: 111.}
}
\description{
Helper function to split data into training and test set
}
