% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{createCVPlot_}
\alias{createCVPlot_}
\title{createCVPlot_}
\usage{
createCVPlot_(cv_obj, filename)
}
\arguments{
\item{filename}{A character string. The filename.}
}
\description{
Helper function to create cross-validation plots
}
