% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_diagnostic.R
\name{createTrainingTest_}
\alias{createTrainingTest_}
\title{createTrainingTest_}
\usage{
createTrainingTest_(diagnosis, mergeset, split = 0.8, seed = 111)
}
\arguments{
\item{diagnosis}{A object. The output of the function `createDiagnosisDesign_()`.}

\item{mergeset}{A merged ExpressionSet.}

\item{split}{A numeric value between 0 and 1. The proportion of the data to be integrated into the training set for machine learning. Default: 0.8.}

\item{seed}{A integer value. Seed to make machine learning algorithms reproducible. Default: 111.}
}
\description{
Helper function to split data into training and test set
}
