% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherfunctions.R
\name{exportDEGannotations_}
\alias{exportDEGannotations_}
\title{exportDEGannotations_}
\usage{
exportDEGannotations_(mergedset)
}
\description{
Helper function to export DEG annotations from mergedset.
}
