% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherfunctions.R
\name{exportDEGannotations_}
\alias{exportDEGannotations_}
\title{exportDEGannotations_}
\usage{
exportDEGannotations_(mergedset, genes)
}
\arguments{
\item{mergedset}{An ExpressionSet. The output of the function `merge_()`.}

\item{genes}{A object. The output of the function `identifyDEGs_()`.}
}
\description{
Helper function to export DEG annotations from mergedset.
}
