% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{extractKEGGterms_}
\alias{extractKEGGterms_}
\title{extractKEGGterms_}
\usage{
extractKEGGterms_(entrez, species)
}
\arguments{
\item{entrez}{A character vector containing the Entrez-IDs.}

\item{species}{A character string indicating the sample's species. Currently supported: "Hs".}
}
\description{
Helper function to extract KEGG terms
}
