% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prognostic_sig.R
\name{generate_expression_pattern}
\alias{generate_expression_pattern}
\title{generate_expression_pattern}
\usage{
generate_expression_pattern(classifier_studies, sig_cov, mergeset,
  study_metadata, sample_metadata, controlname, targetname, targetcol)
}
\arguments{
\item{classifier_studies}{A character vector. Names of the studies
used to train the classifier.}

\item{sig_cov}{A data.frame. Output of the function `univ_cox()`.}

\item{mergeset}{A matrix of merged expression sets (rows = genes, columns = samples).}

\item{study_metadata}{A data frame. The data frame holding the study metadata.}

\item{sample_metadata}{A data frame. The data frame holding the sample metadata.}

\item{controlname}{A character string. Name of the the controls, specified in the 'target' column of `sample_metadata`.}

\item{targetname}{A character string. Name of the the targets, specified
in the 'target' column of `sample_metadata`.}

\item{targetcol}{A character string. Columname of `sample_metadata` holding the targets. Default: "target". Caution: this should not be changed.}
}
\description{
Helper function to generate expression pattern
}
