% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prognostic_sig.R
\name{generateExpressionPattern_}
\alias{generateExpressionPattern_}
\title{generateExpressionPattern_}
\usage{
generateExpressionPattern_(classifier_studies, sigCov, mergeset,
  studyMetadata, sampleMetadata, controlname, targetname, targetcol)
}
\arguments{
\item{classifier_studies}{A character vector. Names of the studies used to train the classifier.}

\item{sigCov}{A data.frame. Output of the function `univCox_()`.}

\item{mergeset}{A matrix of merged expression sets (rows = genes, columns = samples).}

\item{studyMetadata}{A data frame. The data frame holding the study metadata.}

\item{sampleMetadata}{A data frame. The data frame holding the sample metadata.}

\item{controlname}{A character string. Name of the the controls, specified in the 'target' column of `sampleMetadata`.}

\item{targetname}{A character string. Name of the the targets, specified in the 'target' column of `sampleMetadata`.}

\item{targetcol}{A character string. Columname of `sampleMetadata` holding the targets. Default: "target". Caution: this should not be changed.}
}
\description{
Helper function to generate expression pattern
}
