% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_prognostic.R
\name{getSurvivalTime_}
\alias{getSurvivalTime_}
\title{getSurvivalTime_}
\usage{
getSurvivalTime_(studyMetadata, sampleMetadata,
  discoverystudies.w.timedata, targetname, controlname, targetcol, datadir)
}
\arguments{
\item{studyMetadata}{A data frame. The data frame holding the study metadata.}

\item{sampleMetadata}{A data frame. The data frame holding the sample metadata.}

\item{discoverystudies.w.timedata}{A list that contains specifications on the study/studies that contain(s) survival time information.}

\item{targetname}{A character string. Name of the the targets, specified in the 'target' column of `sampleMetadata`.}

\item{controlname}{A character string. Name of the the controls, specified in the 'target' column of `sampleMetadata`.}

\item{targetcol}{A character string. Columname of `sampleMetadata` holding the targets. Default: "target". Caution: this should not be changed.}
}
\description{
Helper function to get survival time.
}
