% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{goDiffReg_}
\alias{goDiffReg_}
\title{goDiffReg_}
\usage{
goDiffReg_(mergeset, design, idtype, entrezids = NULL)
}
\arguments{
\item{mergeset}{A matrix of merged expression sets (rows = genes, columns = samples).}

\item{design}{A object. The output of the function `createDiagnosisDesignBatch_()`.}

\item{entrezids}{A character vector, containing entrez IDs. To be used only if 'idtype'= "affy" (default = NULL).}
}
\description{
Helper function to fitting linear models
}
