% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{goDiffReg_}
\alias{goDiffReg_}
\title{goDiffReg_}
\usage{
goDiffReg_(mergeset, design, idtype, entrezids = NULL)
}
\arguments{
\item{mergeset}{A matrix of merged expression sets (rows = genes, columns = samples).}

\item{design}{A object. The output of the function `createDiagnosisDesign_()`.}

\item{idtype}{A character string. The type of ID used to name the genes. One of 'entrez' or 'affy' intended to use either entrez IDs or
affy IDs. Caution: when using entrez IDs, missing and duplicated IDs are being removed!}

\item{entrezids}{A character vector, containing entrez IDs. To be used only if 'idtype'= "affy" (default = NULL).}
}
\description{
Helper function to fitting linear models
}
