% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{goEnrichmentAnalysis_}
\alias{goEnrichmentAnalysis_}
\title{goEnrichmentAnalysis_}
\usage{
goEnrichmentAnalysis_(gene, OrgDB, organism, fitlm, pathwayid, species,
  plotdir = NULL)
}
\arguments{
\item{gene}{A character vector containing the Entrez-IDs.}

\item{OrgDB}{OrgDB}

\item{organism}{A character string indicating the organism. Currently supported: "hsa".}

\item{fitlm}{An MArrayLM object, returned by `goDiffReg_()`.}

\item{pathwayid}{KEGG pathway ID (like hsa04110 for cell cycle)}

\item{species}{A character string indicating the sample's species. Currently supported: "Hs".}

\item{plotdir}{A character string. Path to the folder to store resulting plots. Default: "./plots/".}
}
\description{
Helper function to perform enrichment analysis
}
