% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG.R
\name{identifyDEGs_}
\alias{identifyDEGs_}
\title{identifyDEGs_}
\usage{
identifyDEGs_(mergeset, design, qValue)
}
\arguments{
\item{mergeset}{A matrix of merged expression sets (rows = genes, columns = samples).}

\item{design}{A object. The output of the function `createDiagnosisDesign_()`.}

\item{qValue}{A numeric value. The output of the function `qSelection_()`.}
}
\description{
Helper function to identify DEGs
}
