% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_knn.R
\name{knn_classifier}
\alias{knn_classifier}
\title{knn_classifier}
\usage{
knn_classifier(
  traininglist = traininglist,
  seed = seed,
  nfolds = nfolds,
  repeats = repeats,
  tunelength = tunelength
)
}
\arguments{
\item{traininglist}{A list object containing the training data. The output
of the function `create_training_test_split()`.}

\item{seed}{Intilization state of random number generator}

\item{nfolds}{A integer. The number of folds used for cross validation
(default: 10).}

\item{repeats}{An integer. The number of repeated cross validations
(default = 5).}

\item{tunelength}{An integer. The caret tuning length (default = 10).}
}
\description{
Helper function to create a knn classifier model
}
