% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prognostic_sig.R
\name{prognosticClassifier_}
\alias{prognosticClassifier_}
\title{prognosticClassifier_}
\usage{
prognosticClassifier_(PatternCom, idtype, validationstudiesinfo, datadir,
  targetcol, targetname, controlname)
}
\arguments{
\item{PatternCom}{A data.frame. The output of the function `generateExpressionPattern_()`.}

\item{validationstudiesinfo}{A list that contains specifications on the study that contains the validation information.}

\item{datadir}{A character string. Path to the data-folder inside the metadata folder.}
}
\description{
Helper function to get survival time.
}
