% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_prognostic.R
\name{prognosticClassifier_}
\alias{prognosticClassifier_}
\title{prognosticClassifier_}
\usage{
prognosticClassifier_(PatternCom, validationstudiesinfo, datatdir,
  targetcol, targetname, controlname)
}
\arguments{
\item{PatternCom}{A data.frame. The output of the function `generateExpressionPattern_()`.}

\item{validationstudiesinfo}{A list that contains specifications on the study that contains the validation information.}

\item{targetcol}{A character string. Columname of `sampleMetadata` holding the targets. Default: "target". Caution: this should not be changed.}

\item{targetname}{A character string. Name of the the targets, specified in the 'target' column of `sampleMetadata`.}

\item{controlname}{A character string. Name of the the controls, specified in the 'target' column of `sampleMetadata`.}
}
\description{
Helper function to get survival time.
}
