% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_rf.R
\name{rf_classifier}
\alias{rf_classifier}
\title{rf_classifier}
\usage{
rf_classifier(traininglist, seed, nfolds, repeats, tunelength)
}
\arguments{
\item{traininglist}{A list object containing the training data. The output
of the function `create_training_test_split()`.}

\item{seed}{Intilization state of random number generator}

\item{nfolds}{A integer. The number of folds used for cross validation
(default: 10).}

\item{repeats}{An integer. The number of repeated cross validations
(default = 5).}

\item{tunelength}{An integer. The caret tuning length (default = 10).}
}
\description{
Helper function to create a rf classifier model
}
