% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigident_signature.R
\name{sigident_signature}
\alias{sigident_signature}
\title{sigident_signature}
\usage{
sigident_signature(
  traininglist,
  type,
  alpha = NULL,
  nfolds = 10,
  repeats = 5,
  tunelength = 10,
  seed
)
}
\arguments{
\item{traininglist}{A list object containing the training data. The output
of the function `create_training_test_split()`.}

\item{type}{A character string. The algorihm used to perform calculations.
Currently implemented are \emph{"elasticnet_grid", "lasso", "elastic",
"svm", "knn"}.}

\item{alpha}{A numeric between 0 and 1. The elastic net mixing parameter
passed to `glmnet::glmnet()`.}

\item{nfolds}{A integer. The number of folds used for cross validation
(default: 10).}

\item{repeats}{An integer. The number of repeated cross validations
(default = 5).}

\item{tunelength}{An integer. The caret tuning length (default = 10).}

\item{seed}{A integer value. Seed to make machine learning algorithms
reproducible. Default: 111.}
}
\description{
Helper function to conduct lasso or elastic net regualization
  for fitting a GLM in order to identify a multi-gene classifier
}
