% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigident.R
\name{sigidentMicroarray}
\alias{sigidentMicroarray}
\title{Perform Signature Analyses in Gene Expression Datasets Derived from MicroArrays}
\usage{
sigidentMicroarray(mergedset, controlname, targetname, studyMetadata,
  sampleMetadata, species, OrgDB, organism, pathwayid,
  deg.q.selection = NULL, seed = 111, traintest.split = 0.8,
  plotdir = "./plots/", csvdir = "./tables/", targetcol = "target")
}
\arguments{
\item{mergedset}{An expression set merged by the method of Hughey and Butte.}
}
\description{

}
\examples{
\dontrun{
}

}
