% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigidentPrognostic.R
\name{sigidentPrognostic}
\alias{sigidentPrognostic}
\title{Perform Prognostic Signature Analyses in Gene Expression
  Datasets Derived from MicroArrays}
\usage{
sigidentPrognostic(
  mergeset,
  sample_metadata,
  idtype,
  genes,
  discoverystudies_w_timedata,
  classifier_studies,
  validationstudiesinfo,
  datadir,
  plotdir = "./plots/",
  csvdir = "./tables/"
)
}
\arguments{
\item{mergeset}{A matrix of merged expression sets (rows = genes,
columns = samples).}

\item{sample_metadata}{A data frame. The data frame holding the
sample metadata.}

\item{idtype}{A character string. The type of ID used to name the
genes. One of 'entrez' or 'affy' intended to use either entrez IDs or
affy IDs. Caution: when using entrez IDs, missing and duplicated IDs
are being removed!}

\item{genes}{A object. The output of the function `identify_degs()`.}

\item{discoverystudies_w_timedata}{A list that contains specifications
on the study/studies that contain(s) survival time information.}

\item{classifier_studies}{A character vector. Names of the studies
used to train the classifier.}

\item{validationstudiesinfo}{A list that contains specifications on
the study that contains the validation information.}

\item{datadir}{A character string. Path to the data-folder inside the
metadata folder.}

\item{plotdir}{A character string. Path to the folder to store resulting
plots. Default: "./plots/".}

\item{csvdir}{A character string. Path to the folder to store output
tables. Default: "./tables/".}
}
\description{
One function to perform prognostic signature analysis.
}
