% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_svm_classifier.R
\name{svm_classifier}
\alias{svm_classifier}
\title{svm_classifier}
\usage{
svm_classifier(traininglist, seed, nfolds)
}
\arguments{
\item{traininglist}{A list object containing the training data. The output
of the function `create_training_test_split()`.}

\item{seed}{Intilization state of random number generator}
}
\description{
Helper function to create a svm classifier model
}
