% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_prognostic.R
\name{univCox_}
\alias{univCox_}
\title{univCox_}
\usage{
univCox_(survtable, entrezids)
}
\arguments{
\item{survtable}{A data.frame. Output of the function `getSurvivalTime_()`.}

\item{entrezids}{A character string. Output of the function `getSurvivalTime_()`.}
}
\description{
Helper function to compute univariate cox regression and determine significance of each gene through separate univariate Cox regressions
}
